requirements.txt
fastapi
uvicorn
langchain
langchain-openai
langchain-community
openai
faiss-cpu>=1.7.4
pymilvus>=2.4.0
unstructured
pdfplumber
tiktoken
python-dotenv
starlette-exporter
prometheus-client
pytest
redis
PyJWT
requests
opentelemetry-sdk
opentelemetry-exporter-otlp
sentry-sdk
rank-bm25
langchain-aws
langchain-classic
opentelemetry-instrumentation-fastapi
